Updates rails to support RubyMine autocomplete
RubyMine complains about not having the rails gem installed, if rails isn't in the gem folder.
As a work around I currently have to install rails to the local user.
Since the only file in the rails folder is README.md it has to be included
for rails to be properly detected by RubyMine.

Closes #61