Handle outdated community ruby-* packages
Quarry uses rubygems index as a source of ruby package versions. It does
not work well when an Arch official package gets stuck at some old version,
e.g. ruby-faraday is at 1.8 for almost a year now when upstream released 2.1
already. This confuses quarry that creates ruby-faraday-1 for version 1.8
and it conflics with official ruby-faraday.

To workaround such situation we use official ruby-* package version as
if it is the latest released gem version. Thus dependency like
'faraday=1.8' gets resolved to ruby-faraday rather than to a slotted
version ruby-faraday-1

Issue #128